algo_trader/requirements.txt
# GUI Framework
PyQt6>=6.4.0

# HTTP Requests
requests>=2.28.0
aiohttp>=3.8.0

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Database
sqlalchemy>=2.0.0

# WebSocket for real-time data
websocket-client>=1.4.0

# Date/Time handling
python-dateutil>=2.8.0

# Encryption for storing credentials
cryptography>=38.0.0

# Configuration
pyyaml>=6.0

# Logging
loguru>=0.6.0

# Async support
asyncio-throttle>=1.0.0

# Plotting for backtesting
matplotlib>=3.6.0

# Technical indicators
ta>=0.10.0
